refactor(desktop): 将会话与 composer 内联组件拆出 App.tsx
Phase 1：提取 tool-call、message-card、composer-surface、layout chrome 等 presentational 组件。
