+++ apps/desktop/src/App.tsx
@@ -240,6 +240,8 @@ import {
   modAltLetterShortcutKbdKeys,
   modLetterShortcutKbdKeys,
   resolveUseMicaBackdrop,
+  isWin32ElectronShell,
+  isDarwinElectronShell,
 } from "@/lib/desktop-shell";
 import { LaunchSplash } from "@/components/launch-splash";
 import { SessionSidebar, type SettingsSidebarTab } from "@/components/session-sidebar";
@@ -276,1592 +278,20 @@ import {
 } from "@/lib/conversation-list-scope";
 
 /** 主会话列最大宽度（居中） */
-const CONVERSATION_MAX_W = "max-w-[min(86vw,44rem)]";
-/** 消息列表、输入区与顶栏横幅共用的水平内边距（侧栏拉宽后列变窄时避免贴窗缘） */
-const CONVERSATION_GUTTER_X = "px-4 sm:px-5";
-const CONVERSATION_GUTTER_NEG_X = "-mx-4 sm:-mx-5";
-/** composer-dock 未量到高度前的滚动床 fallback（原 12rem） */
-const CONVERSATION_COMPOSER_SCROLL_BED_FALLBACK_PX = 192;
-/** 末条消息与 composer 叠层之间可继续下滚的额外留白 */
-const CONVERSATION_SCROLL_BED_EXTRA_PX = 48;
 
-function formatModelPickerLabel(name: string, reasoningEffort: DesktopModelReasoningEffort): string {
-  return `${name} · ${modelReasoningEffortLabel(reasoningEffort)}`;
-}
-
-type ModelPickerItem = DesktopSnapshot["config"]["models"][number];
-
-function ModelPickerRow({
-  model,
-  displayTitle,
-  isActive,
-  onSelectModel,
-}: {
-  model: ModelPickerItem;
-  displayTitle: string;
-  isActive: boolean;
-  onSelectModel: () => void;
-}) {
-  const { getTriggerProps } = useHoverDetailTooltipContext<ModelPickerItem>();
-  const { onPointerEnter, isHighlighted } = getTriggerProps(model);
-
-  return (
-    <HoverDetailTooltip.Anchor itemId={model.name}>
-      <div
-        role="menuitem"
-        tabIndex={-1}
-        className={cn(
-          DESKTOP_OVERLAY_LIST_SUB_TRIGGER,
-          "cursor-pointer outline-none focus:bg-accent focus:text-accent-foreground",
-          (isActive || isHighlighted) && "bg-accent/40",
-        )}
-        onPointerEnter={onPointerEnter}
-        onClick={onSelectModel}
-        onKeyDown={(event) => {
-          if (event.key === "Enter" || event.key === " ") {
-            event.preventDefault();
-            onSelectModel();
-          }
-        }}
-      >
-        <span className={cn(DESKTOP_OVERLAY_LIST_ITEM_PRIMARY, "min-w-0 truncate")}>
-          {displayTitle}
-        </span>
-      </div>
-    </HoverDetailTooltip.Anchor>
-  );
-}
-
-function normalizeSlashPath(value: string): string {
-  return value.replace(/\\/g, "/").replace(/\/+$/g, "");
-}
-
-type EmptyStateWorkspaceSelectorProps = {
-  currentWorkspaceRoot: string;
-  workspaceBinding: DesktopSnapshot["workspaceBinding"];
-  availableWorkspaces: DesktopSnapshot["availableWorkspaces"];
-  disabled?: boolean;
-  onSelectWorkspace(workspaceRoot: string): void;
-  onSelectNoWorkspace(): void;
-  onAddWorkspace(): void;
-};
-
-function EmptyStateWorkspaceSelector({
-  currentWorkspaceRoot,
-  workspaceBinding,
-  availableWorkspaces,
-  disabled,
-  onSelectWorkspace,
-  onSelectNoWorkspace,
-  onAddWorkspace,
-}: EmptyStateWorkspaceSelectorProps) {
-  const { t } = useTranslation();
-  const [workspaceFilter, setWorkspaceFilter] = useState("");
-  const filteredWorkspaces = useMemo(() => {
-    const query = workspaceFilter.trim().toLowerCase();
-    if (!query) {
-      return availableWorkspaces;
-    }
-    return availableWorkspaces.filter((workspace) =>
-      workspace.label.toLowerCase().includes(query) || workspace.path.toLowerCase().includes(query),
-    );
-  }, [availableWorkspaces, workspaceFilter]);
-  const currentWorkspaceLabel = useMemo(
-    () =>
-      resolveWorkspaceSelectorLabel(
-        currentWorkspaceRoot,
-        workspaceBinding,
-        availableWorkspaces,
-        t,
-      ),
-    [availableWorkspaces, currentWorkspaceRoot, t, workspaceBinding],
-  );
-
-  return (
-    <div className="flex justify-start px-0.5">
-      <FilteredOverlayMenu
-        variant="workspace-panel"
-        filterValue={workspaceFilter}
-        onFilterChange={setWorkspaceFilter}
-        filterPlaceholder={t('app.searchWorkspace')}
-        onOpenChange={(open) => {
-          if (!open) {
-            setWorkspaceFilter("");
-          }
-        }}
-        trigger={
-          <FilteredOverlayMenuTrigger asChild>
-            <button
-              type="button"
-              disabled={disabled}
-              aria-label={t('app.selectWorkspace')}
-              className={cn(
-                "inline-flex h-8 max-w-[min(24rem,100%)] min-w-0 items-center gap-1 rounded-md border-0 bg-transparent pr-0.5 pl-1 text-left outline-none hover:bg-muted/40 focus-visible:ring-2 focus-visible:ring-ring/50 disabled:pointer-events-none disabled:opacity-50",
-                instantHoverMotionClass,
-              )}
-            >
-              <span className="min-w-0 flex-1 truncate text-xs text-muted-foreground" title={currentWorkspaceRoot}>
-                {currentWorkspaceLabel}
-              </span>
-              <ChevronDown className="size-3 shrink-0 text-muted-foreground/80" aria-hidden />
-            </button>
-          </FilteredOverlayMenuTrigger>
-        }
-        footer={
-          <>
-            <DropdownMenuItem onSelect={onAddWorkspace} className={cn("gap-1.5", DESKTOP_OVERLAY_LIST_ACTION_ITEM)}>
-              <FolderPlus className="size-3 shrink-0 text-muted-foreground" aria-hidden />
-              <span>{t('app.addWorkspace')}</span>
-            </DropdownMenuItem>
-            <DropdownMenuItem
-              onSelect={onSelectNoWorkspace}
-              className={cn(
-                "gap-1.5",
-                DESKTOP_OVERLAY_LIST_ACTION_ITEM,
-                workspaceBinding === "none" && "bg-accent/40",
-              )}
-            >
-              <MessageSquareText className="size-3 shrink-0 text-muted-foreground" aria-hidden />
-              <span>{t('app.noWorkspace')}</span>
-            </DropdownMenuItem>
-          </>
-        }
-      >
-        {filteredWorkspaces.length === 0 ? (
-          <p className="px-2 py-4 text-center text-xs text-muted-foreground">{t('app.noMatches')}</p>
-        ) : (
-          filteredWorkspaces.map((workspace) => {
-            const selected =
-              workspaceBinding === "project"
-              && sameWorkspacePath(workspace.path, currentWorkspaceRoot);
-            return (
-              <DropdownMenuItem
-                key={workspace.path}
-                onSelect={() => onSelectWorkspace(workspace.path)}
-                className={cn("items-start", DESKTOP_OVERLAY_LIST_ITEM, selected && "bg-accent/40")}
-              >
-                <div className="min-w-0 flex-1">
-                  <div className={DESKTOP_OVERLAY_LIST_ITEM_PRIMARY} title={workspace.label}>
-                    {workspace.label}
-                  </div>
-                  <div className={DESKTOP_OVERLAY_LIST_ITEM_SECONDARY} title={workspace.path}>
-                    {workspace.path}
-                  </div>
-                </div>
-              </DropdownMenuItem>
-            );
-          })
-        )}
-      </FilteredOverlayMenu>
-    </div>
-  );
-}
-
-type ReadLocalImagePreview = (filePath: string) => Promise<string | null>;
-type ReadManagedImagePreview = (reference: string) => Promise<string | null>;
-type ReadLocalVideoPreview = (filePath: string) => Promise<string | null>;
-type ReadManagedVideoPreview = (reference: string) => Promise<string | null>;
-type SaveLocalImageAs = (filePath: string) => Promise<boolean>;
-
-function ToolCallCollapsible({
-  tool,
-  variant = "standalone",
-  readLocalImagePreviewDataUrl,
-  readLocalVideoPreviewUrl,
-  readManagedVideoPreviewUrl,
-  saveLocalImageAs,
-  onOpenSubagentViewer,
-  onAbortShell,
-}: {
-  tool: ToolBlockSnapshot;
-  variant?: "standalone" | "process-nested";
-  readLocalImagePreviewDataUrl: ReadLocalImagePreview;
-  readLocalVideoPreviewUrl: ReadLocalVideoPreview;
-  readManagedVideoPreviewUrl: ReadManagedVideoPreview;
-  saveLocalImageAs: SaveLocalImageAs;
-  onOpenSubagentViewer?: (toolCallId: string) => void;
-  onAbortShell?: (toolCallId: string) => void;
-}) {
-  if (tool.toolName === "finish_task") {
-    return null;
-  }
-
-  if (tool.toolName === "generate_image") {
-    return (
-      <ImageGenerationToolCard
-        tool={tool}
-        readLocalImagePreviewDataUrl={readLocalImagePreviewDataUrl}
-        saveLocalImageAs={saveLocalImageAs}
-      />
-    );
-  }
-
-  if (tool.toolName === "generate_video") {
-    return (
-      <VideoGenerationToolCard
-        tool={tool}
-        readLocalVideoPreviewUrl={readLocalVideoPreviewUrl}
-        readManagedVideoPreviewUrl={readManagedVideoPreviewUrl}
-      />
-    );
-  }
-
-  return (
-    <MinimalToolCallCard
-      tool={tool}
-      variant={variant}
-      onOpenSubagentViewer={onOpenSubagentViewer}
-      onAbortShell={onAbortShell}
-    />
-  );
-}
-
-function ImageGenerationToolCard({
-  tool,
-  readLocalImagePreviewDataUrl,
-  saveLocalImageAs,
-}: {
-  tool: ToolBlockSnapshot;
-  readLocalImagePreviewDataUrl: ReadLocalImagePreview;
-  saveLocalImageAs: SaveLocalImageAs;
-}) {
-  const { t } = useTranslation();
-  const previewableImagePath = tool.imagePaths?.find(isPreviewableImagePath) ?? "";
-  const imagePath = tool.imagePaths?.find((path) => path.trim().length > 0) ?? "";
-  const [previewDataUrl, setPreviewDataUrl] = useState<string | null>(null);
-  const [previewState, setPreviewState] = useState<"idle" | "loading" | "ready" | "unavailable">("idle");
-  const [previewAspectRatio, setPreviewAspectRatio] = useState<number | null>(null);
-  const [viewerOpen, setViewerOpen] = useState(false);
-  const [saving, setSaving] = useState(false);
-
-  useEffect(() => {
-    let cancelled = false;
-    setPreviewDataUrl(null);
-    if (!previewableImagePath) {
-      setPreviewState("unavailable");
-      return () => {
-        cancelled = true;
-      };
-    }
-
-    setPreviewState("loading");
-    void readLocalImagePreviewDataUrl(previewableImagePath)
-      .then((dataUrl) => {
-        if (cancelled) {
-          return;
-        }
-        setPreviewDataUrl(dataUrl);
-        setPreviewState(dataUrl ? "ready" : "unavailable");
-      })
-      .catch(() => {
-        if (!cancelled) {
-          setPreviewState("unavailable");
-        }
-      });
-
-    return () => {
-      cancelled = true;
-    };
-  }, [previewableImagePath, readLocalImagePreviewDataUrl]);
-
-  useEffect(() => {
-    let cancelled = false;
-    setPreviewAspectRatio(null);
-    if (!previewDataUrl) {
-      return () => {
-        cancelled = true;
-      };
-    }
-
-    const image = new Image();
-    image.onload = () => {
-      if (!cancelled && image.naturalWidth > 0 && image.naturalHeight > 0) {
-        setPreviewAspectRatio(image.naturalWidth / image.naturalHeight);
-      }
-    };
-    image.onerror = () => {
-      if (!cancelled) {
-        setPreviewAspectRatio(null);
-      }
-    };
-    image.src = previewDataUrl;
-
-    return () => {
-      cancelled = true;
-      image.onload = null;
-      image.onerror = null;
-    };
-  }, [previewDataUrl]);
-
-  const loading = tool.phase === "preview" || tool.phase === "running" || previewState === "loading";
-  const canInteract = Boolean(previewDataUrl && previewableImagePath);
-  // backdrop-filter 在祖先 opacity 动画期间无法正确合成；卡片 hover 渐显须与 blur 写在同一元素上。
-  const floatingActionButtonClass =
-    "size-8 rounded-full border border-border/50 bg-background/55 text-foreground shadow-sm backdrop-blur-xl transition-[opacity,background-color,border-color,box-shadow] duration-200 ease-out hover:border-border/60 hover:bg-background/72 dark:border-white/12 dark:bg-input/30 dark:hover:bg-input/40 supports-[backdrop-filter]:bg-background/40 dark:supports-[backdrop-filter]:bg-input/25";
-  const floatingActionCardRevealClass =
-    "opacity-0 group-hover/image-card:opacity-100 group-focus-within/image-card:opacity-100";
-  const viewerFrameStyle = previewAspectRatio
-    ? {
-        aspectRatio: String(previewAspectRatio),
-        width: `min(calc(100dvw - 5rem), calc((100dvh - 6rem) * ${previewAspectRatio}), 70rem)`,
-      }
-    : undefined;
-
-  const handleSaveImage = async () => {
-    if (!imagePath || saving) {
-      return;
-    }
-
-    setSaving(true);
-    try {
-      await saveLocalImageAs(imagePath);
-    } finally {
-      setSaving(false);
-    }
-  };
-
-  return (
-    <div className="w-full max-w-[min(28rem,100%)] py-1">
-      <div
-        className={cn(
-          "group/image-card relative aspect-square overflow-hidden rounded-md border border-border/45 bg-muted/20 transition-colors duration-200",
-          canInteract && "cursor-zoom-in hover:border-border/70",
-          tool.phase === "failed" && "border-destructive/45 bg-destructive/5",
-        )}
-        role={canInteract ? "button" : undefined}
-        tabIndex={canInteract ? 0 : undefined}
-        onClick={canInteract ? () => setViewerOpen(true) : undefined}
-        onKeyDown={
-          canInteract
-            ? (event) => {
-                if (event.key === "Enter" || event.key === " ") {
-                  event.preventDefault();
-                  setViewerOpen(true);
-                }
-              }
-            : undefined
-        }
-      >
-        {previewDataUrl ? (
-          <img
-            src={previewDataUrl}
-            alt=""
-            className="size-full object-cover transition-transform duration-300 group-hover/image-card:scale-[1.015]"
-            draggable={false}
-          />
-        ) : (
-          <div className="flex size-full items-center justify-center px-4 text-center">
-            <span
-              className={cn(
-                "text-sm font-medium",
-                loading ? "spirit-thinking-shimmer-text" : "text-muted-foreground",
-              )}
-            >
-              {loading ? t('common.loading') : t('app.previewUnavailable')}
-            </span>
-          </div>
-        )}
-        {previewDataUrl ? (
-          <div className="pointer-events-none absolute inset-0 z-10">
-            <Button
-              type="button"
-              size="icon"
-              variant="ghost"
-              className={cn(
-                "pointer-events-auto absolute bottom-3 left-3",
-                floatingActionButtonClass,
-                floatingActionCardRevealClass,
-              )}
-              onClick={(event) => {
-                event.stopPropagation();
-                void handleSaveImage();
-              }}
-              disabled={saving}
-              title={t('app.downloadImage')}
-              aria-label={t('app.downloadImage')}
-            >
-              {saving ? <LoaderCircle className="size-4 animate-spin" aria-hidden /> : <Download className="size-4" aria-hidden />}
-            </Button>
-            <Button
-              type="button"
-              size="icon"
-              variant="ghost"
-              className={cn(
-                "pointer-events-auto absolute right-3 bottom-3",
-                floatingActionButtonClass,
-                floatingActionCardRevealClass,
-              )}
-              onClick={(event) => {
-                event.stopPropagation();
-                setViewerOpen(true);
-              }}
-              title={t('app.viewLargeImage')}
-              aria-label={t('app.viewLargeImage')}
-            >
-              <Maximize2 className="size-4" aria-hidden />
-            </Button>
-          </div>
-        ) : null}
-      </div>
-      {!previewDataUrl && imagePath ? (
-        <p className="mt-2 truncate font-mono text-[11px] text-muted-foreground" title={imagePath}>
-          {imagePath}
-        </p>
-      ) : null}
-      {tool.phase === "failed" && tool.outputExcerpt ? (
-        <pre className="mt-2 whitespace-pre-wrap rounded-md border border-destructive/20 bg-destructive/5 p-2 font-mono text-xs leading-relaxed text-destructive">
-          {tool.outputExcerpt}
-        </pre>
-      ) : null}
-
-      <Dialog open={viewerOpen} onOpenChange={setViewerOpen}>
-        <DialogContent
-          showCloseButton={false}
-          overlayClassName="bg-background/40 backdrop-blur-md"
-          className="w-auto max-w-none gap-0 border-0 bg-transparent p-0 shadow-none ring-0 sm:max-w-none"
-        >
-          {previewDataUrl ? (
-            <div
-              className="pointer-events-auto relative inline-flex max-h-[calc(100dvh-2rem)] max-w-[calc(100dvw-2rem)] items-center justify-center overflow-hidden rounded-[1.1rem] border border-border/45"
-              style={viewerFrameStyle}
-            >
-              <DialogClose asChild>
-                <Button
-                  type="button"
-                  variant="ghost"
-                  size="icon"
-                  className={cn("absolute top-3 right-3 z-20", floatingActionButtonClass)}
-                  title={t('app.closeImagePreview')}
-                  aria-label={t('app.closeImagePreview')}
-                >
-                  <X className="size-4" aria-hidden />
-                  <span className="sr-only">{t('app.closeImagePreview')}</span>
-                </Button>
-              </DialogClose>
-              <img
-                src={previewDataUrl}
-                alt=""
-                className="block size-full object-contain"
-                draggable={false}
-              />
-              <div className="pointer-events-none absolute top-3 left-3 z-10">
-                <Button
-                  type="button"
-                  size="icon"
-                  variant="ghost"
-                  className={cn("pointer-events-auto", floatingActionButtonClass)}
-                  onClick={() => void handleSaveImage()}
-                  disabled={saving}
-                  title={t('app.downloadImage')}
-                  aria-label={t('app.downloadImage')}
-                >
-                  {saving ? <LoaderCircle className="size-4 animate-spin" aria-hidden /> : <Download className="size-4" aria-hidden />}
-                </Button>
-              </div>
-            </div>
-          ) : null}
-        </DialogContent>
-      </Dialog>
-    </div>
-  );
-}
-
-function VideoGenerationToolCard({
-  tool,
-  readLocalVideoPreviewUrl,
-  readManagedVideoPreviewUrl,
-}: {
-  tool: ToolBlockSnapshot;
-  readLocalVideoPreviewUrl: ReadLocalVideoPreview;
-  readManagedVideoPreviewUrl: ReadManagedVideoPreview;
-}) {
-  const { t } = useTranslation();
-  const videoPath = tool.videoPaths?.find((path) => path.trim().length > 0) ?? "";
-  const previewSourcePath =
-    tool.videoPaths?.find((path) => isManagedGeneratedVideoRef(path) || isPreviewableVideoPath(path)) ?? "";
-  const [previewUrl, setPreviewUrl] = useState<string | null>(null);
-  const [previewState, setPreviewState] = useState<"idle" | "loading" | "ready" | "unavailable">("idle");
-
-  useEffect(() => {
-    let cancelled = false;
-    setPreviewUrl(null);
-    if (!previewSourcePath) {
-      setPreviewState("unavailable");
-      return () => {
-        cancelled = true;
-      };
-    }
-
-    setPreviewState("loading");
-    const resolvePreview = isManagedGeneratedVideoRef(previewSourcePath)
-      ? readManagedVideoPreviewUrl(previewSourcePath)
-      : readLocalVideoPreviewUrl(previewSourcePath);
-
-    void resolvePreview
-      .then((url) => {
-        if (cancelled) {
-          return;
-        }
-        setPreviewUrl(url);
-        setPreviewState(url ? "ready" : "unavailable");
-      })
-      .catch(() => {
-        if (!cancelled) {
-          setPreviewState("unavailable");
-        }
-      });
-
-    return () => {
-      cancelled = true;
-    };
-  }, [previewSourcePath, readLocalVideoPreviewUrl, readManagedVideoPreviewUrl]);
-
-  const loading = tool.phase === "preview" || tool.phase === "running" || previewState === "loading";
-
-  return (
-    <div className="w-full max-w-[min(28rem,100%)] py-1">
-      <div
-        className={cn(
-          "relative aspect-square overflow-hidden rounded-md border border-border/45 bg-muted/20 transition-colors duration-200",
-          tool.phase === "failed" && "border-destructive/45 bg-destructive/5",
-        )}
-      >
-        {previewUrl ? (
-          <video
-            src={previewUrl}
-            className="size-full object-contain"
-            controls
-            preload="metadata"
-          />
-        ) : (
-          <div className="flex size-full items-center justify-center px-4 text-center">
-            <span
-              className={cn(
-                "text-sm font-medium",
-                loading ? "spirit-thinking-shimmer-text" : "text-muted-foreground",
-              )}
-            >
-              {loading ? t('common.loading') : t('app.previewUnavailable')}
-            </span>
-          </div>
-        )}
-      </div>
-      {!previewUrl && videoPath ? (
-        <p className="mt-2 truncate font-mono text-[11px] text-muted-foreground" title={videoPath}>
-          {videoPath}
-        </p>
-      ) : null}
-      {tool.phase === "failed" && tool.outputExcerpt ? (
-        <pre className="mt-2 whitespace-pre-wrap rounded-md border border-destructive/20 bg-destructive/5 p-2 font-mono text-xs leading-relaxed text-destructive">
-          {tool.outputExcerpt}
-        </pre>
-      ) : null}
-    </div>
-  );
-}
-
-type ComposerSurfaceProps = {
-  value: string;
-  localFileAttachments: readonly ComposerLocalFileAttachmentView[];
-  placeholder: string;
-  models: DesktopSnapshot["config"]["models"];
-  catalogHints?: DesktopSnapshot["config"]["modelCatalogHints"];
-  activeModel: string;
-  agentMode: DesktopAgentMode;
-  loopEnabled: boolean;
-  canSend: boolean;
-  canAbort?: boolean;
-  busy: boolean;
-  readOnly?: boolean;
-  onChange(value: string): void;
-  onSubmit(): void;
-  onAbort?(): void;
-  onModelSelect(name: string): void;
-  onModelReasoningEffortSelect(name: string, reasoningEffort: DesktopModelReasoningEffort): void;
-  onAgentModeChange(mode: DesktopAgentMode): void;
-  onLoopEnabledChange?(enabled: boolean): void;
-  richInputRef?: React.RefObject<ComposerRichInputHandle | null>;
-  onKeyDown?(event: ReactKeyboardEvent<HTMLTextAreaElement>): void;
-  onSelectionChange?(selectionStart: number | null): void;
-  showInsertButton?: boolean;
-  canPickLocalFile?: boolean;
-  onInsertWorkspaceFileReferenceTrigger?(): void;
-  onPickLocalFile?(): void | Promise<void>;
-  onInsertSkillTrigger?(): void;
-  onRemoveLocalFileAttachment?(path: string): void;
-  onPaste?(event: ReactClipboardEvent<HTMLTextAreaElement>): void;
-  browserElementAttachments?: readonly BrowserElementAttachment[];
-  onElementAttachmentsChange?(attachments: BrowserElementAttachment[]): void;
-  initialSegments?: readonly RichSegment[] | null;
-  conversationBusy?: boolean;
-  agentModeChipDismissed?: boolean;
-  onAgentModeChipDismissChange?(dismissed: boolean): void;
-};
-
-function ComposerSurface({
-  value,
-  localFileAttachments,
-  placeholder,
-  models,
-  catalogHints,
-  activeModel,
-  agentMode,
-  loopEnabled = false,
-  canSend,
-  canAbort = false,
-  busy,
-  readOnly = false,
-  onChange,
-  onSubmit,
-  onAbort,
-  onModelSelect,
-  onModelReasoningEffortSelect,
-  onAgentModeChange,
-  onLoopEnabledChange,
-  richInputRef,
-  onKeyDown,
-  onSelectionChange,
-  showInsertButton = false,
-  canPickLocalFile = false,
-  onInsertWorkspaceFileReferenceTrigger,
-  onPickLocalFile,
-  onInsertSkillTrigger,
-  onRemoveLocalFileAttachment,
-  onPaste,
-  browserElementAttachments,
-  onElementAttachmentsChange,
-  initialSegments,
-  conversationBusy = false,
-  agentModeChipDismissed = false,
-  onAgentModeChipDismissChange,
-}: ComposerSurfaceProps) {
-  const { t } = useTranslation();
-  const activeModelProfile = useMemo(
-    () => models.find((model) => model.name === activeModel),
-    [activeModel, models],
-  );
-
-  return (
-    <div
-      data-spirit-surface="composer-surface"
-      className="relative overflow-hidden rounded-2xl border border-border/50 bg-background/55 shadow-sm backdrop-blur-xl focus-within:ring-0 hover:border-ring/60 focus-within:border-ring/60 dark:border-white/10 dark:hover:border-white/12 dark:focus-within:border-white/12 dark:bg-input/30 supports-[backdrop-filter]:bg-background/40 dark:supports-[backdrop-filter]:bg-input/25"
-    >
-      <ComposerLocalFileStrip
-        attachments={localFileAttachments}
-        onRemove={(path) => onRemoveLocalFileAttachment?.(path)}
-      />
-      <ComposerRichInput
-        ref={richInputRef}
-        value={value}
-        elementAttachments={browserElementAttachments}
-        initialSegments={initialSegments}
-        placeholder={placeholder}
-        readOnly={readOnly}
-        loopEnabled={loopEnabled}
-        loopChipLabel={t('composer.loopChipLabel')}
-        agentMode={agentMode}
-        planChipLabel={t('composer.planChipLabel')}
-        askChipLabel={t('composer.askChipLabel')}
-        onTextChange={onChange}
-        onElementAttachmentsChange={(atts) => onElementAttachmentsChange?.(atts)}
-        onLoopEnabledChange={onLoopEnabledChange}
-        onAgentModeChange={onAgentModeChange}
-        conversationBusy={conversationBusy}
-        agentModeChipDismissed={agentModeChipDismissed}
-        onAgentModeChipDismissChange={onAgentModeChipDismissChange}
-        onPaste={(e) => onPaste?.(e as unknown as ReactClipboardEvent<HTMLTextAreaElement>)}
-        onKeyDown={(e) => {
-          onKeyDown?.(e as unknown as ReactKeyboardEvent<HTMLTextAreaElement>);
-          if (e.defaultPrevented) return;
-          if (
-            e.key === 'Enter' &&
-            !e.shiftKey &&
-            !e.ctrlKey &&
-            !e.metaKey &&
-            // React synthetic event 的 isComposing 不可靠，必须用 nativeEvent 检测 IME 组合态
-            !e.nativeEvent.isComposing
-          ) {
-            e.preventDefault();
-            if (canSend) onSubmit();
-          }
-        }}
-        onSelectionChange={onSelectionChange}
-      />
-      <div className="flex justify-center px-3 pt-0.5 pb-2">
-        <div className="flex w-full max-w-full items-center justify-between gap-2">
-          <div className="flex min-w-0 flex-1 items-center gap-1.5">
-            {showInsertButton ? (
-              <div className="shrink-0">
-                <ComposerInsertMenu
-                  disabled={readOnly}
-                  canPickLocalFile={canPickLocalFile}
-                  onInsertWorkspaceReference={() => onInsertWorkspaceFileReferenceTrigger?.()}
-                  onPickLocalFile={() => onPickLocalFile?.()}
-                  onInsertSkillTrigger={() => onInsertSkillTrigger?.()}
-                />
-              </div>
-            ) : null}
-            <ModelPickerMenu
-              models={models}
-              catalogHints={catalogHints}
-              activeModelName={activeModel}
-              activeReasoningEffort={activeModelProfile?.reasoningEffort}
-              disabled={readOnly}
-              onModelSelect={onModelSelect}
-              onModelReasoningEffortSelect={onModelReasoningEffortSelect}
-              triggerClassName="max-w-[min(12rem,100%)] pr-0.5 pl-1"
-              menuContentClassName="z-[100]"
-            />
-          </div>
-          {(() => {
-            const hasComposerPayload =
-              value.trim().length > 0 || localFileAttachments.length > 0;
-            const showAbortButton = canAbort && Boolean(onAbort) && !hasComposerPayload;
-            const showEnqueueWhileBusy = canAbort && hasComposerPayload;
-            const sendDisabled = showAbortButton ? false : !canSend || (busy && !canAbort);
-            const actionAriaLabel = showAbortButton
-              ? t("app.abort")
-              : showEnqueueWhileBusy
-                ? t("composer.enqueueWhileBusy")
-                : t("app.send");
-            const actionButton = (
-              <Button
-                type="button"
-                className={cn(
-                  "size-8 shrink-0 rounded-full p-0 shadow-none [&_svg]:size-3.5",
-                  instantHoverMotionClass,
-                )}
-                onClick={showAbortButton ? onAbort : onSubmit}
-                disabled={sendDisabled}
-                aria-label={actionAriaLabel}
-              >
-                {showAbortButton ? (
-                  <Square className="size-3.5" strokeWidth={2.4} aria-hidden />
-                ) : busy ? (
-                  <LoaderCircle className="size-3.5 animate-spin" />
-                ) : (
-                  <ArrowUp className="size-3.5" strokeWidth={2.25} aria-hidden />
-                )}
-              </Button>
-            );
-
-            if (sendDisabled) {
-              return actionButton;
-            }
-
-            return (
-              <Tooltip delayDuration={300}>
-                <TooltipTrigger asChild>{actionButton}</TooltipTrigger>
-                <TooltipContent side="top" sideOffset={4}>
-                  {showAbortButton ? (
-                    <>
-                      {t("app.abort")} <ComposerAbortShortcutKbd />
-                    </>
-                  ) : showEnqueueWhileBusy ? (
-                    t("composer.enqueueWhileBusy")
-                  ) : (
-                    <>
-                      {t("app.send")} <ComposerSendEnterKbd />
-                    </>
-                  )}
-                </TooltipContent>
-              </Tooltip>
-            );
-          })()}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-/** 随 `active` 切换 `.spirit-thinking-shimmer-text`（样式与动画在 `styles.css`）。 */
-function ReasoningLabelWithShimmer({
-  active,
-  activeLabel,
-  idleLabel,
-}: {
-  active: boolean;
-  activeLabel: string;
-  idleLabel: string;
-}) {
-  return (
-    <span
-      className={cn(
-        "shrink-0 text-xs font-medium tracking-wide",
-        active ? "spirit-thinking-shimmer-text" : "text-muted-foreground",
-      )}
-    >
-      {active ? activeLabel : idleLabel}
-    </span>
-  );
-}
-
-function ThinkingLabelWithShimmer({ active }: { active: boolean }) {
-  return (
-    <ReasoningLabelWithShimmer active={active} activeLabel="Thinking" idleLabel="Thought" />
-  );
-}
-
-function CompactionLabelWithShimmer({ active }: { active: boolean }) {
-  return (
-    <ReasoningLabelWithShimmer active={active} activeLabel="Compacting" idleLabel="Compacted" />
-  );
-}
-
-function isLiveStreamingThinkingMessage(
-  message: ConversationMessageSnapshot | undefined,
-  pendingAuxState?: PendingAssistantAux,
-  messages?: readonly ConversationMessageSnapshot[],
-  messageIndex?: number,
-): boolean {
-  return Boolean(
-    message &&
-      isAssistantReasoningLive(message, pendingAuxState, messages, messageIndex),
-  );
-}
-
-function AssistantThinkingCollapsible({
-  message,
-  pendingAuxState,
-  messages,
-  listIndex,
-  collapseDuringToolPreview,
-  readManagedImagePreviewDataUrl,
-  readManagedVideoPreviewUrl,
-}: {
-  message: ConversationMessageSnapshot;
-  pendingAuxState?: PendingAssistantAux;
-  messages: readonly ConversationMessageSnapshot[];
-  listIndex: number;
-  collapseDuringToolPreview: boolean;
-  readManagedImagePreviewDataUrl: ReadManagedImagePreview;
-  readManagedVideoPreviewUrl: ReadManagedVideoPreview;
-}) {
-  const thinking = message.aux?.thinking?.trim() ?? "";
-  const reasoningLive = isAssistantReasoningLive(message, pendingAuxState, messages, listIndex);
-  const showThinkingBody = Boolean(thinking && !isGenericPendingThinkingStatusText(thinking));
-  const thinkingActive = reasoningLive && !collapseDuringToolPreview;
-  if (!thinking && !reasoningLive) {
-    return null;
-  }
-  if (!showThinkingBody && !thinkingActive) {
-    return null;
-  }
-  const autoExpanded = thinkingActive && showThinkingBody;
-  const [manualOpen, setManualOpen] = useState(false);
-  const prevAutoExpandedRef = useRef(autoExpanded);
-
-  useEffect(() => {
-    if (prevAutoExpandedRef.current && !autoExpanded) {
-      setManualOpen(false);
-    }
-    prevAutoExpandedRef.current = autoExpanded;
-  }, [autoExpanded]);
-
-  const expanded = autoExpanded || manualOpen;
-  const interactive = !autoExpanded;
-
-  return (
-    <AnimatedCollapse
-      open={expanded}
-      onOpenChange={(open) => {
-        if (!interactive) {
-          return;
-        }
-        setManualOpen(open);
-      }}
-      className="min-w-0 py-0.5"
-    >
-      <AnimatedCollapseTrigger
-        className={cn(
-          "group flex w-full min-w-0 items-center gap-1 text-left outline-none",
-          interactive ? "cursor-pointer focus-visible:ring-2 focus-visible:ring-ring/50" : "cursor-default",
-        )}
-      >
-        <ThinkingLabelWithShimmer active={thinkingActive} />
-        {interactive ? (
-          <ChevronRight
-            className={cn(
-              "size-3 shrink-0 text-muted-foreground/55 transition-all duration-150",
-              "opacity-0 group-hover:opacity-100 group-focus-visible:opacity-100",
-              expanded && "rotate-90",
-            )}
-            aria-hidden
-          />
-        ) : null}
-      </AnimatedCollapseTrigger>
-      {showThinkingBody ? (
-        <AnimatedCollapseContent className="min-w-0">
-          <div className="overflow-hidden pt-1.5 [&_p:last-child]:mb-0 [&_ul:last-child]:mb-0 [&_ol:last-child]:mb-0 [&_blockquote:last-child]:mb-0 [&_pre:last-child]:mb-0">
-            <AgentMarkdownMessage
-              content={thinking}
-              streaming={thinkingActive && expanded}
-              tone="muted"
-              readManagedImagePreviewDataUrl={readManagedImagePreviewDataUrl}
-              readManagedVideoPreviewUrl={readManagedVideoPreviewUrl}
-            />
-          </div>
-        </AnimatedCollapseContent>
-      ) : null}
-    </AnimatedCollapse>
-  );
-}
-
-function AssistantCompactionCollapsible({
-  message,
-  pendingAuxState,
-  readManagedImagePreviewDataUrl,
-  readManagedVideoPreviewUrl,
-}: {
-  message: ConversationMessageSnapshot;
-  pendingAuxState?: PendingAssistantAux;
-  readManagedImagePreviewDataUrl: ReadManagedImagePreview;
-  readManagedVideoPreviewUrl: ReadManagedVideoPreview;
-}) {
-  const compaction = message.aux?.compaction?.trim() ?? "";
-  const compactionLive = assistantCompactionLive(message, pendingAuxState);
-  const showCompactionBody = Boolean(
-    compaction && !isGenericPendingCompactionStatusText(compaction),
-  );
-  const compactionActive = compactionLive;
-  const autoExpanded = compactionActive && showCompactionBody;
-  const [manualOpen, setManualOpen] = useState(false);
-  const prevAutoExpandedRef = useRef(autoExpanded);
-
-  useEffect(() => {
-    if (prevAutoExpandedRef.current && !autoExpanded) {
-      setManualOpen(false);
-    }
-    prevAutoExpandedRef.current = autoExpanded;
-  }, [autoExpanded]);
-
-  const expanded = autoExpanded || manualOpen;
-  const interactive = !autoExpanded;
-
-  return (
-    <AnimatedCollapse
-      open={expanded}
-      onOpenChange={(open) => {
-        if (!interactive) {
-          return;
-        }
-        setManualOpen(open);
-      }}
-      className="min-w-0 py-0.5"
-    >
-      <AnimatedCollapseTrigger
-        className={cn(
-          "group flex w-full min-w-0 items-center gap-1 text-left outline-none",
-          interactive ? "cursor-pointer focus-visible:ring-2 focus-visible:ring-ring/50" : "cursor-default",
-        )}
-      >
-        <CompactionLabelWithShimmer active={compactionActive} />
-        {interactive ? (
-          <ChevronRight
-            className={cn(
-              "size-3 shrink-0 text-muted-foreground/55 transition-all duration-150",
-              "opacity-0 group-hover:opacity-100 group-focus-visible:opacity-100",
-              expanded && "rotate-90",
-            )}
-            aria-hidden
-          />
-        ) : null}
-      </AnimatedCollapseTrigger>
-      {showCompactionBody ? (
-        <AnimatedCollapseContent className="min-w-0">
-          <div className="overflow-hidden pt-1.5 [&_p:last-child]:mb-0 [&_ul:last-child]:mb-0 [&_ol:last-child]:mb-0 [&_blockquote:last-child]:mb-0 [&_pre:last-child]:mb-0">
-            <AgentMarkdownMessage
-              content={compaction}
-              streaming={compactionActive}
-              tone="muted"
-              readManagedImagePreviewDataUrl={readManagedImagePreviewDataUrl}
-              readManagedVideoPreviewUrl={readManagedVideoPreviewUrl}
-            />
-          </div>
-        </AnimatedCollapseContent>
-      ) : null}
-    </AnimatedCollapse>
-  );
-}
-
-function MessageCard({
-  composerSessionKey,
-  conversationListScopeKey,
-  messages,
-  message,
-  listIndex,
-  compactAfterPrevious,
-  tightenAfterPreviousMeta,
-  showContinueButton,
-  continueTarget,
-  continueBusy,
-  rewindText,
-  rewindLocalFileAttachments,
-  rewindBrowserElementAttachments,
-  rewindSelected,
-  rewindCanSubmit,
-  rewindBusy,
-  rewindRichInputRef,
-  onRewindElementAttachmentsChange,
-  canPickLocalFile,
-  models,
-  catalogHints,
-  activeModel,
-  agentMode,
-  onContinue,
-  onRewindChange,
-  onRewindStart,
-  onRewindSubmit,
-  onRewindRemoveLocalFileAttachment,
-  onRewindPickLocalFile,
-  onRewindPaste,
-  onModelSelect,
-  onModelReasoningEffortSelect,
-  onAgentModeChange,
-  pendingAuxState,
-  readManagedImagePreviewDataUrl,
-  readManagedVideoPreviewUrl,
-  readLocalImagePreviewDataUrl,
-  readLocalVideoPreviewUrl,
-  saveLocalImageAs,
-  onOpenSubagentViewer,
-  onAbortShell,
-  queuedCanMoveUp = false,
-  queueActionBusy = false,
-  onQueueMoveUp,
-  onQueueSendNow,
-  onQueueDelete,
-  hiddenByProcessGroup = false,
-}: {
-  composerSessionKey: string;
-  conversationListScopeKey: string;
-  messages: readonly ConversationMessageSnapshot[];
-  pendingAuxState?: PendingAssistantAux;
-  message: ConversationMessageSnapshot;
-  listIndex: number;
-  hiddenByProcessGroup?: boolean;
-  compactAfterPrevious: boolean;
-  tightenAfterPreviousMeta: boolean;
-  showContinueButton: boolean;
-  continueTarget?: ConversationMessageSnapshot;
-  continueBusy: boolean;
-  rewindText: string;
-  rewindLocalFileAttachments: readonly ComposerLocalFileAttachmentView[];
-  rewindBrowserElementAttachments: readonly BrowserElementAttachment[];
-  rewindSelected: boolean;
-  rewindCanSubmit: boolean;
-  rewindBusy: boolean;
-  rewindRichInputRef: React.RefObject<ComposerRichInputHandle | null>;
-  onRewindElementAttachmentsChange(attachments: BrowserElementAttachment[]): void;
-  canPickLocalFile: boolean;
-  models: DesktopSnapshot["config"]["models"];
-  catalogHints?: DesktopSnapshot["config"]["modelCatalogHints"];
-  activeModel: string;
-  agentMode: DesktopAgentMode;
-  onContinue(message: ConversationMessageSnapshot): void;
-  onRewindChange(value: string): void;
-  onRewindStart(message: ConversationMessageSnapshot, listIndex: number): void;
-  onRewindSubmit(): void;
-  onRewindRemoveLocalFileAttachment(path: string): void;
-  onRewindPickLocalFile(): void;
-  onRewindPaste(event: ReactClipboardEvent<HTMLTextAreaElement>): void;
-  onModelSelect(name: string): void;
-  onModelReasoningEffortSelect(name: string, reasoningEffort: DesktopModelReasoningEffort): void;
-  onAgentModeChange(mode: DesktopAgentMode): void;
-  readManagedImagePreviewDataUrl: ReadManagedImagePreview;
-  readManagedVideoPreviewUrl: ReadManagedVideoPreview;
-  readLocalImagePreviewDataUrl: ReadLocalImagePreview;
-  readLocalVideoPreviewUrl: ReadLocalVideoPreview;
-  saveLocalImageAs: SaveLocalImageAs;
-  onOpenSubagentViewer?: (toolCallId: string) => void;
-  onAbortShell?: (toolCallId: string) => void;
-  queuedCanMoveUp?: boolean;
-  queueActionBusy?: boolean;
-  onQueueMoveUp?(queueId: string): void;
-  onQueueSendNow?(queueId: string): void;
-  onQueueDelete?(queueId: string): void;
-}) {
-  const { t } = useTranslation();
-  const isUser = message.role === "user";
-  const isQueuedUser = isUser && message.queued === true && typeof message.queueId === "string";
-  const canStartRewind =
-    isUser && message.canRewind === true && !message.pending && message.queued !== true;
-  const userBubble =
-    "rounded-2xl rounded-br-md border border-border/50 bg-muted px-3 py-2.5 shadow-sm";
-  const subagentStatusSurface =
-    !isUser && message.content.trim() ? isSubagentStatusSurfaceMessage(message) : false;
-  const showThinkingCollapsible =
-    !hiddenByProcessGroup &&
-    shouldShowAssistantThinkingCollapsible(message, pendingAuxState, messages, listIndex);
-  const showCompactionCollapsible =
-    !hiddenByProcessGroup &&
-    shouldShowAssistantCompactionCollapsible(message, pendingAuxState);
-  const collapseThinkingDuringToolPreview = shouldCollapseThinkingDuringToolPreview(
-    messages,
-    listIndex,
-  );
-  const rewindInitialSegments = useMemo(
-    () =>
-      rewindSelected
-        ? messageContentToRichSegments(message.content, String(message.id))
-        : null,
-    [rewindSelected, message.content, message.id],
-  );
-  return (
-    <div
-      id={conversationMessageStableId(message, composerSessionKey, conversationListScopeKey)}
-      data-spirit-surface="message-row"
-      data-spirit-message-role={message.role}
-      data-spirit-message-pending={message.pending ? "true" : "false"}
-      className={cn(
-        "scroll-mt-4 flex w-full pb-3 last:pb-0",
-        compactAfterPrevious && "-mt-4",
-        tightenAfterPreviousMeta && "-mt-3",
-        isUser ? "justify-end" : "justify-start",
-        rewindSelected && "relative z-40",
-      )}
-    >
-      <div
-        data-spirit-surface={isUser ? "message-user" : "message-assistant"}
-        className={cn(
-          "min-w-0 space-y-2",
-          isUser
-            ? rewindSelected
-              ? "ml-auto w-full max-w-[min(100%,36rem)]"
-              : "max-w-[min(72%,22rem)]"
-            : "w-full",
-        )}
-      >
-        {rewindSelected && isUser ? (
-          <ComposerSurface
-            key={`rewind-composer-${message.id}`}
-            richInputRef={rewindRichInputRef}
-            value={rewindText}
-            initialSegments={rewindInitialSegments}
-            browserElementAttachments={rewindBrowserElementAttachments}
-            onElementAttachmentsChange={onRewindElementAttachmentsChange}
-            localFileAttachments={rewindLocalFileAttachments}
-            onChange={onRewindChange}
-            onSubmit={onRewindSubmit}
-            placeholder={t('app.typeMessage')}
-            models={models}
-            catalogHints={catalogHints}
-            activeModel={activeModel}
-            agentMode={agentMode}
-            loopEnabled={false}
-            onModelSelect={onModelSelect}
-            onModelReasoningEffortSelect={onModelReasoningEffortSelect}
-            onAgentModeChange={onAgentModeChange}
-            onLoopEnabledChange={() => {}}
-            canSend={rewindCanSubmit}
-            busy={rewindBusy}
-            showInsertButton
-            canPickLocalFile={canPickLocalFile}
-            onPickLocalFile={onRewindPickLocalFile}
-            onRemoveLocalFileAttachment={onRewindRemoveLocalFileAttachment}
-            onPaste={onRewindPaste}
-          />
-        ) : null}
-        {showThinkingCollapsible ? (
-          <AssistantThinkingCollapsible
-            message={message}
-            pendingAuxState={pendingAuxState}
-            messages={messages}
-            listIndex={listIndex}
-            collapseDuringToolPreview={collapseThinkingDuringToolPreview}
-            readManagedImagePreviewDataUrl={readManagedImagePreviewDataUrl}
-            readManagedVideoPreviewUrl={readManagedVideoPreviewUrl}
-          />
-        ) : null}
-        {showCompactionCollapsible ? (
-          <AssistantCompactionCollapsible
-            message={message}
-            pendingAuxState={pendingAuxState}
-            readManagedImagePreviewDataUrl={readManagedImagePreviewDataUrl}
-            readManagedVideoPreviewUrl={readManagedVideoPreviewUrl}
-          />
-        ) : null}
-        {isUser && !rewindSelected ? (
-          isQueuedUser && message.queueId && onQueueMoveUp && onQueueSendNow && onQueueDelete ? (
-            <QueuedUserMessageHoverActions
-              queueId={message.queueId}
-              canMoveUp={queuedCanMoveUp}
-              busy={queueActionBusy}
-              onMoveUp={onQueueMoveUp}
-              onSendNow={onQueueSendNow}
-              onDelete={onQueueDelete}
-            >
-              <UserMessageBubble
-                message={message}
-                userBubbleClassName={userBubble}
-                canStartRewind={false}
-                queued
-                onRewindStart={() => onRewindStart(message, listIndex)}
-                readLocalImagePreviewDataUrl={readLocalImagePreviewDataUrl}
-              />
-            </QueuedUserMessageHoverActions>
-          ) : (
-            <UserMessageBubble
-              message={message}
-              userBubbleClassName={userBubble}
-              canStartRewind={canStartRewind}
-              queued={message.queued === true}
-              onRewindStart={() => onRewindStart(message, listIndex)}
-              readLocalImagePreviewDataUrl={readLocalImagePreviewDataUrl}
-            />
-          )
-        ) : null}
-        {!isUser && message.content.trim() ? (
-          subagentStatusSurface ? (
-            <p className="text-sm leading-relaxed text-muted-foreground">{message.content}</p>
-          ) : (
-          <div data-spirit-surface="message-bubble">
-            <AgentMarkdownMessage
-              content={message.content}
-              streaming={message.pending}
-              className="font-sans"
-              readManagedImagePreviewDataUrl={readManagedImagePreviewDataUrl}
-              readManagedVideoPreviewUrl={readManagedVideoPreviewUrl}
-            />
-          </div>
-          )
-        ) : null}
-        {!isUser && message.aux?.finishTaskNotice ? (
-          <p className="text-xs leading-relaxed text-muted-foreground">
-            {message.aux.finishTaskNotice}
-          </p>
-        ) : null}
-        {!isUser && message.tool ? (
-          <ToolCallCollapsible
-            tool={message.tool}
-            readLocalImagePreviewDataUrl={readLocalImagePreviewDataUrl}
-            readLocalVideoPreviewUrl={readLocalVideoPreviewUrl}
-            readManagedVideoPreviewUrl={readManagedVideoPreviewUrl}
-            saveLocalImageAs={saveLocalImageAs}
-            onOpenSubagentViewer={onOpenSubagentViewer}
-            onAbortShell={onAbortShell}
-          />
-        ) : null}
-        {!isUser && showContinueButton && continueTarget ? (
-          <div className="ml-auto flex max-w-[min(72%,22rem)] justify-end pt-1">
-            <Button
-              type="button"
-              variant="outline"
-              size="sm"
-              className="h-8 px-4"
-              onClick={() => onContinue(continueTarget)}
-              disabled={continueBusy}
-            >
-              {t('app.continue')}
-            </Button>
-          </div>
-        ) : null}
-      </div>
-    </div>
-  );
-}
-
-function isGrayMetaLineMessage(message: ConversationMessageSnapshot | undefined): boolean {
-  return isGrayMetaLeadingMessage(message) && isGrayMetaTrailingMessage(message);
-}
-
-/** Windows Electron：使用 `titleBarOverlay` + 自绘顶栏；macOS 仍走系统菜单栏 */
-function isWin32ElectronShell(): boolean {
-  if (!isElectronChrome() || typeof navigator === "undefined") {
-    return false;
-  }
-  return /Windows/i.test(navigator.userAgent);
-}
-
-/** macOS Electron：`titleBarStyle: hiddenInset`，需预留红绿灯安全区 */
-function isDarwinElectronShell(): boolean {
-  if (!isElectronChrome()) {
-    return false;
-  }
-  return window.spiritDesktop?.platform === "darwin";
-}
-
-function WebHostPairingGate({
-  busy,
-  error,
-  onPair,
-}: {
-  busy: boolean;
-  error: string;
-  onPair(code: string): Promise<boolean>;
-}) {
-  const { t } = useTranslation();
-  const [code, setCode] = useState("");
-  const [localError, setLocalError] = useState("");
-
-  const submit = () => {
-    const normalized = code.trim();
-    if (!normalized) {
-      setLocalError(t('app.enterPairingCode'));
-      return;
-    }
-    void onPair(normalized).then((ok) => {
-      if (!ok) {
-        setLocalError(t('app.pairingFailed'));
-      }
-    });
-  };
-
-  return (
-    <div className="flex h-[100dvh] items-center justify-center bg-background px-4 text-foreground">
-      <Card className="w-full max-w-sm rounded-lg">
-        <CardHeader>
-          <CardTitle>{t('app.firstTimePairing')}</CardTitle>
-          <CardDescription>{t('app.pairingDescription')}</CardDescription>
-        </CardHeader>
-        <CardContent className="space-y-4">
-          <div className="grid gap-2">
-            <Label htmlFor="web-host-pairing-code">{t('app.pairingCode')}</Label>
-            <Input
-              id="web-host-pairing-code"
-              value={code}
-              inputMode="numeric"
-              autoComplete="one-time-code"
-              onChange={(event) => {
-                setLocalError("");
-                setCode(event.target.value);
-              }}
-              onKeyDown={(event) => {
-                if (event.key === "Enter" && !busy) {
-                  submit();
-                }
-              }}
-            />
-          </div>
-          {localError || (error && !error.includes('需要完成首次配对')) ? (
-            <p className="text-sm text-destructive">{localError || error}</p>
-          ) : null}
-          <Button type="button" className="w-full" disabled={busy} onClick={submit}>
-            {busy ? <LoaderCircle className="size-4 animate-spin" /> : null}
-            {t('app.pair')}
-          </Button>
-        </CardContent>
-      </Card>
-    </div>
-  );
-}
-
-function SessionSidebarShortcutKbd() {
-  const keys = modLetterShortcutKbdKeys("B");
-
-  return (
-    <KbdGroup>
-      {isMacDesktopPlatform() ? (
-        keys.map((key) => <Kbd key={key}>{key}</Kbd>)
-      ) : (
-        <>
-          <Kbd>Ctrl</Kbd>
-          <span>+</span>
-          <Kbd>B</Kbd>
-        </>
-      )}
-    </KbdGroup>
-  );
-}
-
-function NewSessionShortcutKbd() {
-  const keys = modLetterShortcutKbdKeys("N");
-
-  return (
-    <KbdGroup>
-      {isMacDesktopPlatform() ? (
-        keys.map((key) => <Kbd key={key}>{key}</Kbd>)
-      ) : (
-        <>
-          <Kbd>Ctrl</Kbd>
-          <span>+</span>
-          <Kbd>N</Kbd>
-        </>
-      )}
-    </KbdGroup>
-  );
-}
-
-function ComposerSendEnterKbd() {
-  const { t } = useTranslation();
-
-  return (
-    <Kbd aria-label={t("composer.sendEnterKey")}>
-      <CornerDownLeft className="size-3" aria-hidden />
-    </Kbd>
-  );
-}
-
-function ComposerAbortShortcutKbd() {
-  const keys = ctrlLetterShortcutKbdKeys("C");
-
-  return (
-    <KbdGroup>
-      {isMacDesktopPlatform() ? (
-        keys.map((key) => <Kbd key={key}>{key}</Kbd>)
-      ) : (
-        <>
-          <Kbd>Ctrl</Kbd>
-          <span>+</span>
-          <Kbd>C</Kbd>
-        </>
-      )}
-    </KbdGroup>
-  );
-}
-
-function WorkspaceToolsShortcutKbd() {
-  const keys = modAltLetterShortcutKbdKeys("B");
-
-  return (
-    <KbdGroup>
-      {isMacDesktopPlatform() ? (
-        keys.map((key) => <Kbd key={key}>{key}</Kbd>)
-      ) : (
-        <>
-          <Kbd>Ctrl</Kbd>
-          <span>+</span>
-          <Kbd>Alt</Kbd>
-          <span>+</span>
-          <Kbd>B</Kbd>
-        </>
-      )}
-    </KbdGroup>
-  );
-}
-
-function DesktopLayoutChromeBar({
-  useMicaBackdrop,
-  showWorkspaceToggle,
-  workspaceToolsOpen = false,
-  onToggleWorkspaceTools,
-  sessionTitle,
-  subagentPromptText,
-  onExitSubagentViewer,
-  onNewSession,
-  newSessionBusy = false,
-}: {
-  useMicaBackdrop: boolean;
-  showWorkspaceToggle: boolean;
-  workspaceToolsOpen?: boolean;
-  onToggleWorkspaceTools?: () => void;
-  sessionTitle?: string | null;
-  subagentPromptText?: string | null;
-  onExitSubagentViewer?: () => void;
-  onNewSession?: () => void;
-  newSessionBusy?: boolean;
-}) {
-  const { t } = useTranslation();
-  const { open: sessionSidebarOpen, toggle: onToggleSessionSidebar } = useSessionSidebarChrome();
-  const showTrailingActions = showWorkspaceToggle;
-  const trimmedSessionTitle = sessionTitle?.trim() ?? "";
-  return (
-    <div
-      role="toolbar"
-      aria-label={t('app.sidebarAndTools')}
-      data-spirit-surface="layout-chrome"
-      data-session-sidebar-open={sessionSidebarOpen ? "true" : "false"}
-      className={cn(
-        "flex h-8 shrink-0 items-center gap-2 px-1.5",
-        showTrailingActions ? "justify-between" : "justify-start",
-        desktopMicaTintClass(useMicaBackdrop),
-      )}
-    >
-      <div className="flex min-w-0 items-center">
-        <Tooltip delayDuration={300}>
-          <TooltipTrigger asChild>
-            <Button
-              type="button"
-              variant="ghost"
-              size="icon"
-              className={cn(DESKTOP_CHROME_TOGGLE_ICON_BTN, "mr-1")}
-              onClick={onToggleSessionSidebar}
-              aria-label={sessionSidebarOpen ? t('app.hideSidebar') : t('app.showSidebar')}
-              aria-expanded={sessionSidebarOpen}
-              {...(sessionSidebarOpen ? { "aria-controls": "session-sidebar-panel" } : {})}
-            >
-              {sessionSidebarOpen ? <PanelLeftClose className="size-3.5" aria-hidden /> : <PanelLeftOpen className="size-3.5" aria-hidden />}
-            </Button>
-          </TooltipTrigger>
-          <TooltipContent side="top" sideOffset={4}>
-            {sessionSidebarOpen ? t("app.hideSidebar") : t("app.showSidebar")}{" "}
-            <SessionSidebarShortcutKbd />
-          </TooltipContent>
-        </Tooltip>
-        {onNewSession ? (
-          <div
-            className={cn(
-              "shrink-0 overflow-hidden",
-              DESKTOP_SHELL_LAYOUT_TRANSITION,
-              sessionSidebarOpen
-                ? "pointer-events-none mr-0 w-0 opacity-0"
-                : "mr-1 w-7 opacity-100",
-            )}
-            aria-hidden={sessionSidebarOpen}
-          >
-            <Tooltip delayDuration={300}>
-              <TooltipTrigger asChild>
-                <Button
-                  type="button"
-                  variant="ghost"
-                  size="icon"
-                  className={DESKTOP_CHROME_TOGGLE_ICON_BTN}
-                  onClick={onNewSession}
-                  disabled={newSessionBusy}
-                  tabIndex={sessionSidebarOpen ? -1 : undefined}
-                  aria-label={t("sidebar.newSession")}
-                >
-                  <Plus className="size-3.5" aria-hidden />
-                </Button>
-              </TooltipTrigger>
-              <TooltipContent side="top" sideOffset={4}>
-                {t("sidebar.newSession")}{" "}
-                <NewSessionShortcutKbd />
-              </TooltipContent>
-            </Tooltip>
-          </div>
-        ) : null}
-        {trimmedSessionTitle ? (
-          <SessionChromeBreadcrumb
-            sessionTitle={trimmedSessionTitle}
-            subagentPromptText={subagentPromptText}
-            onExitSubagentViewer={onExitSubagentViewer}
-          />
-        ) : null}
-      </div>
-      {showTrailingActions ? (
-        <div className="flex items-center gap-1">
-          {showWorkspaceToggle ? (
-            <Tooltip delayDuration={300}>
-              <TooltipTrigger asChild>
-                <Button
-                  type="button"
-                  variant="ghost"
-                  size="icon"
-                  className={DESKTOP_CHROME_TOGGLE_ICON_BTN}
-                  onClick={() => onToggleWorkspaceTools?.()}
-                  aria-label={workspaceToolsOpen ? t('app.collapseTools') : t('app.expandTools')}
-                  aria-expanded={workspaceToolsOpen}
-                  {...(workspaceToolsOpen ? { "aria-controls": "workspace-tools-panel" } : {})}
-                >
-                  {workspaceToolsOpen ? (
-                    <PanelRightClose className="size-3.5" aria-hidden />
-                  ) : (
-                    <PanelRightOpen className="size-3.5" aria-hidden />
-                  )}
-                </Button>
-              </TooltipTrigger>
-              <TooltipContent side="top" sideOffset={4}>
-                {workspaceToolsOpen ? t("app.collapseTools") : t("app.expandTools")}{" "}
-                <WorkspaceToolsShortcutKbd />
-              </TooltipContent>
-            </Tooltip>
-          ) : null}
-        </div>
-      ) : null}
-    </div>
-  );
-}
+import {
+  CONVERSATION_COMPOSER_SCROLL_BED_FALLBACK_PX,
+  CONVERSATION_GUTTER_NEG_X,
+  CONVERSATION_GUTTER_X,
+  CONVERSATION_MAX_W,
+  CONVERSATION_SCROLL_BED_EXTRA_PX,
+} from "@/lib/conversation-layout-constants";
+import { ComposerSurface } from "@/components/composer/composer-surface";
+import { MessageCard } from "@/components/conversation/message-card";
+import { ToolCallCollapsible } from "@/components/tool-call/tool-call-collapsible";
+import { EmptyStateWorkspaceSelector } from "@/components/empty-state-workspace-selector";
+import { WebHostPairingGate } from "@/components/web-host-pairing-gate";
+import { DesktopLayoutChromeBar } from "@/components/layout/desktop-layout-chrome-bar";
 
 export default function App() {
   const { t, i18n } = useTranslation();
