+++ apps/desktop/src/components/composer/composer-shortcut-kbds.tsx
@@ -1,0 +1,33 @@
+import { CornerDownLeft } from "lucide-react";
+import { useTranslation } from "react-i18next";
+
+import { Kbd, KbdGroup } from "@/components/ui/kbd";
+import { ctrlLetterShortcutKbdKeys, isMacDesktopPlatform } from "@/lib/desktop-shell";
+
+export function ComposerSendEnterKbd() {
+  const { t } = useTranslation();
+
+  return (
+    <Kbd aria-label={t("composer.sendEnterKey")}>
+      <CornerDownLeft className="size-3" aria-hidden />
+    </Kbd>
+  );
+}
+
+export function ComposerAbortShortcutKbd() {
+  const keys = ctrlLetterShortcutKbdKeys("C");
+
+  return (
+    <KbdGroup>
+      {isMacDesktopPlatform() ? (
+        keys.map((key) => <Kbd key={key}>{key}</Kbd>)
+      ) : (
+        <>
+          <Kbd>Ctrl</Kbd>
+          <span>+</span>
+          <Kbd>C</Kbd>
+        </>
+      )}
+    </KbdGroup>
+  );
+}
