+++ apps/desktop/src/lib/conversation-layout-constants.ts
@@ -1,0 +1,9 @@
+/** 主会话列最大宽度（居中） */
+export const CONVERSATION_MAX_W = "max-w-[min(86vw,44rem)]";
+/** 消息列表、输入区与顶栏横幅共用的水平内边距（侧栏拉宽后列变窄时避免贴窗缘） */
+export const CONVERSATION_GUTTER_X = "px-4 sm:px-5";
+export const CONVERSATION_GUTTER_NEG_X = "-mx-4 sm:-mx-5";
+/** composer-dock 未量到高度前的滚动床 fallback（原 12rem） */
+export const CONVERSATION_COMPOSER_SCROLL_BED_FALLBACK_PX = 192;
+/** 末条消息与 composer 叠层之间可继续下滚的额外留白 */
+export const CONVERSATION_SCROLL_BED_EXTRA_PX = 48;
